+++ presets/vitest.config.ts
@@ -1,0 +1,12 @@
+import vue from '@vitejs/plugin-vue';
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  plugins: [vue()],
+  esbuild: { jsx: 'automatic' },
+  test: {
+    environment: 'happy-dom',
+    globals: true,
+    include: ['**/*.test.{ts,tsx}'],
+  },
+});